primitives.rs から arrays.rs と memory.rs を切り出す
- src/primitives/arrays.rs: 配列操作系プリミティブ関数を移動
  - to_array_prim, from_array_prim, array_prim, array_get_prim,
    array_addr_prim, array_len_prim, array_concat_prim,
    check_array_element_value (pub(super))
- src/primitives/memory.rs: メモリアクセス系プリミティブ関数を移動
  - fetch_prim, store_prim, set_prim,
    array_pool_idx_from_cell, is_global_array_pool_idx,
    promote_array_pool_idx_to_global, check_dict_reference_write,
    write_array_element
- write_array_element 内の check_array_element_value 呼び出しを
  super::arrays::check_array_element_value(...) に変更
- primitives.rs に mod arrays / mod memory と pub use を追加
- 既存 API パス (crate::primitives::*) は維持、挙動変更なし
