+++ src/primitives.rs
@@ -12,185 +12,21 @@ use std::time::{SystemTime, UNIX_EPOCH};
 // `primitives.rs` remains the façade and registration entry point; the
 // `pub use` re-exports keep `crate::primitives::<name>` paths working for
 // existing callers and tests.
+mod arrays;
 mod compare;
 mod logic;
+mod memory;
 mod numeric;
 mod stack;
 mod strings;
 
+pub use arrays::*;
 pub use compare::*;
 pub use logic::*;
+pub use memory::*;
 pub use numeric::*;
 pub use stack::*;
 pub use strings::*;
-
-/// FETCH — fetch a value from an address and push it onto the stack.
-pub fn fetch_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let addr = vm.pop()?;
-    match addr {
-        Cell::DictAddr(a) => {
-            let value = vm.dict_read(a)?;
-            vm.push(value)?;
-            Ok(())
-        }
-        Cell::StackAddr(a) => {
-            let value = vm.local_read(a)?;
-            vm.push(value)?;
-            Ok(())
-        }
-        Cell::ArrayAddr { pool_idx, elem_idx } => {
-            let arr = vm.arrays.get(pool_idx).ok_or(TbxError::IndexOutOfBounds {
-                index: pool_idx,
-                size: vm.arrays.len(),
-            })?;
-            let size = arr.len();
-            if elem_idx >= size {
-                return Err(TbxError::ArrayIndexOutOfBounds {
-                    index: elem_idx as i64,
-                    size,
-                });
-            }
-            let value = arr[elem_idx].clone();
-            vm.push(value)?;
-            Ok(())
-        }
-        _ => Err(TbxError::TypeError {
-            expected: "address",
-            got: "non-address",
-        }),
-    }
-}
-
-/// Extract the array pool index from a `Cell::Array`.
-///
-/// Returns `None` for any cell that is not `Cell::Array`.  `Cell::Str` is
-/// `Rc<str>`-backed and has no VM-managed pool index.
-fn array_pool_idx_from_cell(cell: &Cell) -> Option<usize> {
-    match cell {
-        Cell::Array(idx) => Some(*idx),
-        _ => None,
-    }
-}
-
-fn is_global_array_pool_idx(vm: &VM, pool_idx: usize) -> bool {
-    pool_idx < vm.global_array_pool_len
-}
-
-fn promote_array_pool_idx_to_global(vm: &mut VM, pool_idx: usize) {
-    vm.global_array_pool_len = vm.global_array_pool_len.max(pool_idx + 1);
-}
-
-fn check_dict_reference_write(vm: &mut VM, value: &Cell) -> Result<(), TbxError> {
-    let Some(pool_idx) = array_pool_idx_from_cell(value) else {
-        return Ok(());
-    };
-
-    if is_global_array_pool_idx(vm, pool_idx) {
-        return Ok(());
-    }
-
-    if vm.is_executing_top_level() {
-        promote_array_pool_idx_to_global(vm, pool_idx);
-        return Ok(());
-    }
-
-    Err(TbxError::ArrayFrameEscape)
-}
-
-/// STORE — pop addr (top) then value (below), and store value at addr.
-///
-/// Stack convention: `[..., value, addr]` → STORE → `[...]`
-pub fn store_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let addr = vm.pop()?;
-    let value = vm.pop()?;
-    match addr {
-        Cell::DictAddr(a) => {
-            check_dict_reference_write(vm, &value)?;
-            vm.dict_write_at(a, value)?;
-            Ok(())
-        }
-        Cell::StackAddr(a) => {
-            vm.local_write(a, value)?;
-            Ok(())
-        }
-        Cell::ArrayAddr { pool_idx, elem_idx } => {
-            write_array_element(vm, pool_idx, elem_idx, value)
-        }
-        _ => Err(TbxError::TypeError {
-            expected: "address",
-            got: "non-address",
-        }),
-    }
-}
-
-/// SET — pop value (top) then addr (below), and store value at addr.
-///
-/// Designed for the `SET &var, value` statement pattern where `&var` is
-/// pushed before `value` (left-to-right argument evaluation via comma).
-/// Stack convention: `[..., addr, value]` → SET → `[...]`
-pub fn set_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let value = vm.pop()?;
-    let addr = vm.pop()?;
-    match addr {
-        Cell::DictAddr(a) => {
-            check_dict_reference_write(vm, &value)?;
-            vm.dict_write_at(a, value)?;
-            Ok(())
-        }
-        Cell::StackAddr(a) => {
-            vm.local_write(a, value)?;
-            Ok(())
-        }
-        Cell::ArrayAddr { pool_idx, elem_idx } => {
-            write_array_element(vm, pool_idx, elem_idx, value)
-        }
-        _ => Err(TbxError::TypeError {
-            expected: "address",
-            got: "non-address",
-        }),
-    }
-}
-
-/// Check that `value` is a permitted array element value.
-///
-/// `Cell::Array` is always rejected (nested arrays are not supported).
-/// `Cell::Str(Rc<str>)` is allowed: the `Rc` handle keeps the string alive
-/// independently of any stack frame, so there is no dangling risk (#591).
-/// All other scalar types are accepted unconditionally.
-fn check_array_element_value(value: &Cell) -> Result<(), TbxError> {
-    match value {
-        Cell::Array(_) => Err(TbxError::InvalidArrayElement { got: "Array" }),
-        _ => Ok(()),
-    }
-}
-
-/// Write `value` to element `elem_idx` of the array at `pool_idx`.
-fn write_array_element(
-    vm: &mut VM,
-    pool_idx: usize,
-    elem_idx: usize,
-    value: Cell,
-) -> Result<(), TbxError> {
-    // Validate before get_mut() to avoid borrow conflicts.
-    check_array_element_value(&value)?;
-    let pool_size = vm.arrays.len();
-    let arr = vm
-        .arrays
-        .get_mut(pool_idx)
-        .ok_or(TbxError::IndexOutOfBounds {
-            index: pool_idx,
-            size: pool_size,
-        })?;
-    let size = arr.len();
-    if elem_idx >= size {
-        return Err(TbxError::ArrayIndexOutOfBounds {
-            index: elem_idx as i64,
-            size,
-        });
-    }
-    arr[elem_idx] = value;
-    Ok(())
-}
 
 /// GET_OUTPUT — take the current output buffer and return it as a `Cell::Str`.
 /// The output buffer is cleared after this call.
@@ -985,252 +821,6 @@ fn emit_jump_target_to_dict(vm: &mut VM, label_n: i64) -> Result<(), TbxError> {
             .patch_list
             .push((label_n, patch_pos));
     }
-    Ok(())
-}
-
-/// TO_ARRAY — collect n values from the stack into an array.
-///
-/// The compiler emits `LIT Int(n)` before the Xt for variadic primitives, so the
-/// arity is on top of the stack when this function runs.
-///
-/// Stack before call: `[arg0, arg1, ..., arg(n-1), Int(n)]`
-/// Stack after call:  `[Cell::Array(pool_idx)]`
-///
-/// The returned `Cell::Array` is bound to the current frame and must not escape.
-/// TO_ARRAY with zero arguments (`TO_ARRAY()`) produces an empty array.
-pub fn to_array_prim(vm: &mut VM) -> Result<(), TbxError> {
-    // Pop the arity pushed by the compiler.
-    let n = vm.pop_int()?;
-    if n < 0 {
-        return Err(TbxError::InvalidArgument {
-            message: format!("TO_ARRAY arity must be non-negative, got {n}"),
-        });
-    }
-    let count = n as usize;
-    // Pop `count` values in reverse order, then reverse to restore original order.
-    let mut elems: Vec<Cell> = Vec::with_capacity(count);
-    for _ in 0..count {
-        let elem = vm.pop()?;
-        // Reject nested arrays; Cell::Str(Rc<str>) is now allowed (#591).
-        check_array_element_value(&elem)?;
-        elems.push(elem);
-    }
-    elems.reverse();
-    let pool_idx = vm.arrays.len();
-    vm.arrays.push(elems);
-    vm.push(Cell::Array(pool_idx))?;
-    Ok(())
-}
-
-/// FROM_ARRAY — expand an array onto the stack.
-///
-/// Pops `Cell::Array(pool_idx)` from the stack and pushes every element of the
-/// array onto the stack in order (index 0 first).
-///
-/// Stack before call: `[Cell::Array(pool_idx)]`
-/// Stack after call:  `[elem0, elem1, ..., elem(n-1)]`
-pub fn from_array_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let pool_idx = match vm.pop()? {
-        Cell::Array(idx) => idx,
-        other => {
-            return Err(TbxError::TypeError {
-                expected: "Array",
-                got: other.type_name(),
-            })
-        }
-    };
-    let elems = vm
-        .arrays
-        .get(pool_idx)
-        .ok_or(TbxError::IndexOutOfBounds {
-            index: pool_idx,
-            size: vm.arrays.len(),
-        })?
-        .clone();
-    for elem in elems {
-        vm.push(elem)?;
-    }
-    Ok(())
-}
-
-/// ARRAY — create an array of N elements and push its handle onto the stack.
-///
-/// Pops `Cell::Int(n)` from the stack (n > 0), pushes `n` `Cell::None` elements
-/// into `vm.arrays`, and pushes `Cell::Array(pool_idx)` as the handle.
-///
-/// Arrays created inside a word are bound to that stack frame and freed automatically
-/// when the owning word returns (EXIT/RETURN_VAL truncates the pool).
-pub fn array_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let n = vm.pop_int()?;
-    if n <= 0 {
-        return Err(TbxError::InvalidArgument {
-            message: format!("ARRAY size must be positive, got {n}"),
-        });
-    }
-    let size = n as usize;
-    let idx = vm.arrays.len();
-    vm.arrays.push(vec![Cell::None; size]);
-    vm.push(Cell::Array(idx))?;
-    Ok(())
-}
-
-/// ARRAY_GET — read an element from an array.
-///
-/// Stack: `[..., Cell::Array(pool_idx), Cell::Int(elem_idx)]` → `value`
-///
-/// Array indices are 1-based from the user's perspective: valid range is `1..=N`.
-/// The index is translated to 0-based internally before accessing the Vec.
-pub fn array_get_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let elem_idx_raw = vm.pop_int()?;
-    let pool_idx = match vm.pop()? {
-        Cell::Array(idx) => idx,
-        other => {
-            return Err(TbxError::TypeError {
-                expected: "Array",
-                got: other.type_name(),
-            })
-        }
-    };
-    let arr = vm.arrays.get(pool_idx).ok_or(TbxError::IndexOutOfBounds {
-        index: pool_idx,
-        size: vm.arrays.len(),
-    })?;
-    let size = arr.len();
-    // Translate 1-based user index to 0-based internal index.
-    // Index 0 or negative is out of bounds.
-    if elem_idx_raw < 1 {
-        return Err(TbxError::ArrayIndexOutOfBounds {
-            index: elem_idx_raw,
-            size,
-        });
-    }
-    let elem_idx = (elem_idx_raw - 1) as usize;
-    if elem_idx >= size {
-        return Err(TbxError::ArrayIndexOutOfBounds {
-            index: elem_idx_raw,
-            size,
-        });
-    }
-    let value = arr[elem_idx].clone();
-    vm.push(value)?;
-    Ok(())
-}
-
-/// ARRAY_ADDR — compute the address of an array element.
-///
-/// Stack: `[..., Cell::Array(pool_idx), Cell::Int(elem_idx)]` → `Cell::ArrayAddr { pool_idx, elem_idx }`
-///
-/// Array indices are 1-based from the user's perspective: valid range is `1..=N`.
-/// The index is translated to 0-based internally before storing in `Cell::ArrayAddr`.
-pub fn array_addr_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let elem_idx_raw = vm.pop_int()?;
-    let pool_idx = match vm.pop()? {
-        Cell::Array(idx) => idx,
-        other => {
-            return Err(TbxError::TypeError {
-                expected: "Array",
-                got: other.type_name(),
-            })
-        }
-    };
-    // Validate bounds at address-computation time.
-    let arr = vm.arrays.get(pool_idx).ok_or(TbxError::IndexOutOfBounds {
-        index: pool_idx,
-        size: vm.arrays.len(),
-    })?;
-    let size = arr.len();
-    // Translate 1-based user index to 0-based internal index.
-    // Index 0 or negative is out of bounds.
-    if elem_idx_raw < 1 {
-        return Err(TbxError::ArrayIndexOutOfBounds {
-            index: elem_idx_raw,
-            size,
-        });
-    }
-    let elem_idx = (elem_idx_raw - 1) as usize;
-    if elem_idx >= size {
-        return Err(TbxError::ArrayIndexOutOfBounds {
-            index: elem_idx_raw,
-            size,
-        });
-    }
-    vm.push(Cell::ArrayAddr { pool_idx, elem_idx })?;
-    Ok(())
-}
-
-/// ARRAY_LEN — return the length of an array.
-///
-/// Pops `Cell::Array(pool_idx)` from the stack and pushes the number of elements
-/// as `Cell::Int`.
-///
-/// Stack: `[..., Cell::Array(pool_idx)]` → `Cell::Int(len)`
-pub fn array_len_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let pool_idx = match vm.pop()? {
-        Cell::Array(idx) => idx,
-        other => {
-            return Err(TbxError::TypeError {
-                expected: "Array",
-                got: other.type_name(),
-            })
-        }
-    };
-    let arr = vm.arrays.get(pool_idx).ok_or(TbxError::IndexOutOfBounds {
-        index: pool_idx,
-        size: vm.arrays.len(),
-    })?;
-    let len = arr.len() as i64;
-    vm.push(Cell::Int(len))?;
-    Ok(())
-}
-
-/// ARRAY_CONCAT — concatenate two arrays and return a new array.
-///
-/// Pops two `Cell::Array` handles from the stack and pushes a new array whose
-/// contents are all elements of `a` followed by all elements of `b`.
-///
-/// Stack: `[..., a: Array, b: Array]` → `Cell::Array(new_idx)`
-pub fn array_concat_prim(vm: &mut VM) -> Result<(), TbxError> {
-    let b_pool_idx = match vm.pop()? {
-        Cell::Array(idx) => idx,
-        other => {
-            return Err(TbxError::TypeError {
-                expected: "Array",
-                got: other.type_name(),
-            })
-        }
-    };
-    let a_pool_idx = match vm.pop()? {
-        Cell::Array(idx) => idx,
-        other => {
-            return Err(TbxError::TypeError {
-                expected: "Array",
-                got: other.type_name(),
-            })
-        }
-    };
-
-    let a = vm
-        .arrays
-        .get(a_pool_idx)
-        .ok_or(TbxError::IndexOutOfBounds {
-            index: a_pool_idx,
-            size: vm.arrays.len(),
-        })?;
-    let b = vm
-        .arrays
-        .get(b_pool_idx)
-        .ok_or(TbxError::IndexOutOfBounds {
-            index: b_pool_idx,
-            size: vm.arrays.len(),
-        })?;
-
-    let mut result: Vec<Cell> = Vec::with_capacity(a.len() + b.len());
-    result.extend_from_slice(a);
-    result.extend_from_slice(b);
-
-    let pool_idx = vm.arrays.len();
-    vm.arrays.push(result);
-    vm.push(Cell::Array(pool_idx))?;
     Ok(())
 }
 
@@ -3274,34 +2864,6 @@ mod tests {
         vm.push(Cell::DictAddr(0)).unwrap();
         store_prim(&mut vm).unwrap();
         assert_eq!(vm.dict_read(0).unwrap(), Cell::string("hello"));
-    }
-
-    #[test]
-    fn test_array_pool_idx_from_cell_array_only() {
-        // array_pool_idx_from_cell returns Some only for Cell::Array;
-        // Cell::Str is Rc<str>-backed and has no array pool index.
-        assert_eq!(array_pool_idx_from_cell(&Cell::Array(3)), Some(3));
-        assert_eq!(array_pool_idx_from_cell(&Cell::string("hello")), None);
-        assert_eq!(
-            array_pool_idx_from_cell(&Cell::ArrayAddr {
-                pool_idx: 1,
-                elem_idx: 2,
-            }),
-            None
-        );
-        assert_eq!(array_pool_idx_from_cell(&Cell::DictAddr(0)), None);
-        assert_eq!(array_pool_idx_from_cell(&Cell::StackAddr(0)), None);
-    }
-
-    #[test]
-    fn test_promote_array_pool_idx_to_global_never_moves_boundary_backward() {
-        let mut vm = VM::new();
-
-        vm.global_array_pool_len = 5;
-        promote_array_pool_idx_to_global(&mut vm, 1);
-        assert_eq!(vm.global_array_pool_len, 5);
-        promote_array_pool_idx_to_global(&mut vm, 7);
-        assert_eq!(vm.global_array_pool_len, 8);
     }
 
     // --- PUTCHR tests ---
